refactor: change damage property in sword config to float

+++ appFabric/1.20.1/src/main/java/net/kenddie/fantasyweapons/config/FWConfig.java
@@ -27,7 +27,7 @@ public class FWConfig {
     @Configuration
     public static class SwordConfig {
         public static final SwordConfig DEFAULT = new SwordConfig();
-        public int damage = 3;
+        public float damage = 3;
         public float speed = -2.4f;
     }
 }
